+++ examples/resources/uptime_statuspage_component/import.sh
@@ -1,0 +1,2 @@
+# Import using composite ID: statuspage_id:component_id
+terraform import uptime_statuspage_component.api 123:456