+++ examples/resources/uptime_statuspage_incident/import.sh
@@ -1,0 +1,2 @@
+# Import using composite ID: statuspage_id:incident_id
+terraform import uptime_statuspage_incident.outage 123:789