+++ examples/resources/uptime_statuspage_subscriber/import.sh
@@ -1,0 +1,2 @@
+# Import using composite ID: statuspage_id:subscriber_id
+terraform import uptime_statuspage_subscriber.email 123:456