+++ examples/resources/uptime_statuspage_subscription_domain_block/import.sh
@@ -1,0 +1,2 @@
+# Import using composite ID: statuspage_id:domain_id
+terraform import uptime_statuspage_subscription_domain_block.spam 123:456